added support for "context" and "priority"
Add support for the "context" and "priority" arguments in the
"add_meta_box" function. These arguments have been added to the
"register_post_options_section" method, which are then passed on to the
call to the "add_meta_box" function. This update allows the user to have
more control over the placement of the meta boxes.